train/pyproject.toml
[project]
name = "manipulator-lerobot"
version = "0.1.0"
description = "LeRobot data collection and training for manipulator toy-sorting"
requires-python = "==3.12.*"
dependencies = [
    "lerobot>=0.5.0",
    "pyzmq>=26.0.0",
    "msgpack>=1.0.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
